feat: Initial commit
Initial commit for neighborhood map project. So far, I have the Google
map kinda working with geolocation and the ability to change cities via
an input box. Plan to get this working with eventful.com's API to find
events happening on the same day as this is being accessed to then place them on the map.